arm64: dts: sunxi: OrangePi PC 2: Add Ethernet support
Add the required DT magic to enable Ethernet on the board.
This includes the regulator (which is controlled by a GPIO pin) and
the actual Ethernet MAC node, referring the RGMII pins of the device.

This is lifted from the respective Orange Pi Plus patch by
Corentin Labbe <clabbe.montjoie@gmail.com>.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>