clk: sunxi-ng: support R40 SoC
Allwinner R40 SoC have a clock controller module in the style of the
SoCs beyond sun6i, however, it's more rich and complex.

Add support for it.

As the user manual is still not available, this driver is made with
knowledge from clk-sun8iw11.c in the BSP kernel source.