ARM: dts: sun8i: add device tree for Lichee Pi Zero with Dock
Lichee Pi Zero features a dock, which adds some functions, and should be
soldered with the core board of Lichee Pi Zero.

Add support for the dock as a new dts, as soldering is needed to make
the dock usable and there's functions enabled on the dock that is
unavailable at the 2.54mm pins of the core board.

Currently no new feature on core board is available in the mainline
kernel, so the dock dt file only includes the core board device tree.

Signed-off-by: Icenowy Zheng <icenowy@aosc.xyz>