drm: sun4i: do not check downstream panel/bridge when probing TCON
Sometimes the downstream panel/bridge will need to be probed after TCON
is probed; however the TCON probing code will check for the downstream
panel/bridge, so it becomes a dead lock -- both components are waiting
for the probe of another one.

As the pipeline will check the presence of all components, remove the
check in TCON code.

Signed-off-by: Icenowy Zheng <icenowy@aosc.io>