irqchip/gic: Support broken always secure implementations
Some GIC implementations (at least the ones on Allwinner A80/A83T SoCs)
remain in secure state even if the CPU is in non-secure state.

Add support for such broken always secure GIC implementations.

When a GIC implementation is tagged as broken and the CPU is under
non-secure state, the alias registers will be accessed instead of the
common ones.

Signed-off-by: Chen-Yu Tsai <wens@csie.org>
Signed-off-by: Icenowy Zheng <icenowy@aosc.io>