drm: bridge: analogix: fix AUX address-only read
When doing address-only read on AUX I2C, the size of the message is 0,
and no data is needed to be written to the bridge's registers. However,
currently the code still calls regmap_bulk_write with a count of 0,
which will lead to failure.

Only write the data to the bridge's registers via regmap_bulk_write when
the message's size is not 0.

Signed-off-by: Icenowy Zheng <icenowy@aosc.io>