mfd: axp20x: add support for power key on AXP806
When AXP806 is working in "Self-work mode" (which means no other PMIC is
present), a power key can be connected to the EN/PWRON pin. In master or
slave mode the pin will work as EN and there's no power key, so nothing
will be triggered.

Add support for the power key on AXP806.

Signed-off-by: Icenowy Zheng <icenowy@aosc.io>