drm/bridge/synopsys: dw-hdmi: re-run dw_hdmi_setup when setting mode
Currently dw_hdmi_setup is only run when the dw-hdmi bridge is enabled,
with the mode set last time.

When the bridge is enabled before any mode is set (this may happen when
booting), the mode won't be set at all, some setup steps will be
skipped or fail, and the HDMI output may not work.

Re-run dw_hdmi_setup when setting mode, in order to prevent such
situation.

Signed-off-by: Icenowy Zheng <icenowy@aosc.io>