dt-bindings: clock: sun8i-de2: add H6 DE3 CCU compatible string
The Allwinner H6 SoC comes with "Allwinner DE3", which is a modified
version of DE2. From the perspective of clock, a ROT clock is added, but
the other clock keeps the same.

Add compatible string for the DE3 CCU. The ROT clock slice number will
be added when introducing the H6 DE3 CCU support.

Signed-off-by: Icenowy Zheng <icenowy@aosc.io>