arm64: allwinner: a64: dts: add volume keys support for PineTab
PineTab has a pair of volumn keys (up and down) which are connected via
a resistor ladder on the LRADC pin of the A64 SoC. Volume + key is
active at 0.2V and Volume - is at 0.4V according to calculation.

Add volume keys support for PineTab by enabling the LRADC and fill key
voltage infomation.

Signed-off-by: Icenowy Zheng <icenowy@aosc.io>