drm/bridge: analogix-anx6345: fix acquisition of the regulators
When calling regulator_get(), the regulator name should not contain
"-supply" (it will be attached to the regulator name to construct the DT
property name by regulator framework). However, the driver currently
passed bogus "-supply", which breaks the support to its DT binding.

Fix this by remove the bogus "-supply" in regulator names.

Fixes: 6aa192698089 ("drm/bridge: Add Analogix anx6345 support")
Signed-off-by: Icenowy Zheng <icenowy@aosc.io>