arm64: rockchip: dts: rk3399: disable cdn_dp on Pinebook Pro
The corresponding extcon support for fusb302 is not yet available, and
thus the Type-C controller is not tied with the DP controller. However
the DP controller do not support such configuration, and will fail to
probe, and prevent the full display pipeline from being probed.

Disable cdn_dp controller now for Pinebook Pro, in order to make at
least internal eDP panel work.

Signed-off-by: Icenowy Zheng <icenowy@aosc.io>